Move ChunkSetMetadata into the chunking package
Bertan's change, and the first step of the vector_db refactor rather than
part of the chunking work. The model describes a chunk set, so it belongs
beside Chunk rather than in vector_db, and index_documents no longer
reaches into vector_db to import it.

index_documents' VectorDatabase import is repointed at the retrieval
subpackage in the same hunk, following the removal of the top-level
re-exports.
